% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metsta.R
\name{make_metsta}
\alias{make_metsta}
\title{Creation of the 'metsta' dataset}
\usage{
make_metsta(path = ".", save_it = FALSE)
}
\arguments{
\item{path}{character. Data once created are saved in the folder 
\code{file.path(path, "data")}.}

\item{save_it}{logical. If \code{TRUE}, the result is saved as an \code{.RData} file in the 
folder \code{file.path(path, "data")}.}
}
\value{
Returns invisibly the tibble created.
}
\description{
The function \code{make_metsta} creates the \code{\link[metsyn]{metsta}} 
dataset from the file \code{postesSynop.csv} downloaded 
\href{https://donneespubliques.meteofrance.fr/?fond=produit&id_produit=90&id_rubrique=32}{here}.
}
\examples{
\dontrun{
dir.create("data-raw")
make_metsta(save_it = TRUE)
}

}
\seealso{
\code{\link[metsyn]{metsta}}, 
\code{\link[metsyn]{make_metsyn}}
}

