% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metdes.R
\docType{data}
\name{metdes}
\alias{metdes}
\title{Descriptive information on the 'metsyn' dataset}
\format{a \code{\link[tibble]{tibble}} with 5 columns and 59 rows.}
\source{
Meteo France, see \href{https://donneespubliques.meteofrance.fr/?fond=produit&id_produit=90&id_rubrique=32}{here}.
}
\description{
The \code{metdes} dataset gives some descriptive information on the 
\code{\link[metsyn]{metsyn}} dataset. 

The dataset contains the following columns: 
\itemize{
  \item \code{Short_Name}: short name of the variable, in French; 
  \item \code{Long_name_French}: name of the variable, in French; 
  \item \code{Long_Name_English}: currently not provided yet; 
  \item \code{Type}: type of the variable (one of \code{character}, 
  \code{numeric}, \code{integer}); 
  \item \code{Unit}: physical unit of the variable. 
}
}
\note{
This dataset is distributed by Meteo France under the terms of the 
\href{https://www.etalab.gouv.fr/wp-content/uploads/2014/05/Open_Licence.pdf}{Open Licence 1.0}, 
provided by \href{https://www.etalab.gouv.fr/en/qui-sommes-nous}{Etalab} and 
designed to be compatible with the "Creative Commons Attribution 2.0" 
(CC-BY 2.0) licence of Creative Commons. 
Etalab is the task force under the French Prime Minister's authority 
leading Open Government Data policy for France.
}
\keyword{datasets}

