% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metsta.R
\docType{data}
\name{metsta}
\alias{metsta}
\title{Meteo France Synop Meteorogical Stations}
\format{a \code{\link[tibble]{tibble}} with 5 columns and 62 rows.}
\source{
Meteo France, see \href{https://donneespubliques.meteofrance.fr/?fond=produit&id_produit=90&id_rubrique=32}{here}.
}
\description{
This dataset contains the following columns: 
\itemize{
  \item \code{Id}: WMO meteorological station id; 
  \item \code{Name}: name of the meteorological station; 
  \item \code{Latitude}, \code{Longitude}, \code{Altitude}: coordinates of 
  the meteorological station. 
}
}
\note{
This dataset is distributed by Meteo France under the terms of the 
\href{https://www.etalab.gouv.fr/wp-content/uploads/2014/05/Open_Licence.pdf}{Open Licence 1.0}, 
provided by \href{https://www.etalab.gouv.fr/en/qui-sommes-nous}{Etalab} and 
designed to be compatible with the "Creative Commons Attribution 2.0" 
(CC-BY 2.0) licence of Creative Commons. 
Etalab is the task force under the French Prime Minister's authority 
leading Open Government Data policy for France.
}
\keyword{datasets}

