% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metsyn.R
\docType{data}
\name{metsyn}
\alias{metsyn}
\title{Meteo France Synop Dataset (1996-2016)}
\format{a \code{\link[tibble]{tibble}} with 59 columns and 3516433 rows.}
\source{
Meteo France, see \href{https://donneespubliques.meteofrance.fr/?fond=produit&id_produit=90&id_rubrique=32}{here}.
}
\description{
This dataset is made of meteorological data recorded every three hours from 
1996 to 2016 on 62 French meteorological stations. 

The columns contained in this dataset are described by the 
\code{\link[metsyn]{metdes}} dataset.
}
\note{
This dataset is distributed by Meteo France under the terms of the 
\href{https://www.etalab.gouv.fr/wp-content/uploads/2014/05/Open_Licence.pdf}{Open Licence 1.0}, 
provided by \href{https://www.etalab.gouv.fr/en/qui-sommes-nous}{Etalab} and 
designed to be compatible with the "Creative Commons Attribution 2.0" 
(CC-BY 2.0) licence of Creative Commons. 
Etalab is the task force under the French Prime Minister's authority 
leading Open Government Data policy for France.
}
\seealso{
\code{\link[metsyn]{metdes}} for the dataset which gives some descriptive 
information on \code{metsyn}; 

\code{\link[metsyn]{metsta}} for the dataset on the meteorological stations 
involved.
}
\keyword{datasets}

