% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metdes.R
\name{make_metdes}
\alias{make_metdes}
\title{Creation of the 'metdes' dataset}
\usage{
make_metdes(path = ".", save_it = FALSE)
}
\arguments{
\item{path}{character. Data once created are saved in the folder 
\code{file.path(path, "data")}.}

\item{save_it}{logical. If \code{TRUE}, the result is saved as an \code{.RData} file in the 
folder \code{file.path(path, "data")}.}
}
\value{
Returns invisibly the tibble created.
}
\description{
The function \code{make_metdes} creates the \code{\link[metsyn]{metdes}} 
dataset, which contains descriptive information on the 
\code{\link[metsyn]{metsyn}} dataset.
}
\examples{
\dontrun{
dir.create("data-raw")
make_metdes(save_it = TRUE)
}

}
\seealso{
\code{\link[metsyn]{metdes}}, 
\code{\link[metsyn]{make_metsyn}}.
}

