% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metsyn.R
\name{make_metsyn}
\alias{make_metsyn}
\title{Creation of the 'metsyn' dataset}
\usage{
make_metsyn(path = ".", save_it = FALSE)
}
\arguments{
\item{path}{character. Data once created are saved in the folder 
\code{file.path(path, "data")}.}

\item{save_it}{logical. If \code{TRUE}, the result is saved as an \code{.RData} file in the 
folder \code{file.path(path, "data")}.}
}
\value{
Returns invisibly the tibble created.
}
\description{
The function \code{make_metsyn} creates the \code{\link[metsyn]{metsyn}} 
dataset from the files downloaded with 
\code{\link[metsyn]{download_monthly_synop}}.
}
\seealso{
\code{\link[metsyn]{metsyn}}, 
\code{\link[metsyn]{download_monthly_synop}}
}

